Changed monkey weight to a float instead of an int.
git-svn-id: svn+ssh://gleek.ecs.baylor.edu/matrr/trunk@78 8ba6b188-5121-491e-b013-2137287ea23c

+++ matrr/models.py
@@ -128,8 +128,8 @@ class Monkey(models.Model):
   mky_birthdate = models.CharField('Date of Birth', blank=True, null=True,
                                    max_length=20,
                                    help_text='Please enter the date the monkey was born on.')
-  mky_weight = models.IntegerField('Weight', blank=True, null=True,
-                                   help_text='The weight of the monkey.  This should be the weight at time of necropsy (or a recent weight if the necropsy has not yet occurred).')
+  mky_weight = models.FloatField('Weight', blank=True, null=True,
+                                 help_text='The weight of the monkey.  This should be the weight at time of necropsy (or a recent weight if the necropsy has not yet occurred).')
   mky_drinking = models.BooleanField('Drinking', null=False,
                                      help_text='Was this monkey given alcohol?')
   mky_housing_control = models.BooleanField('Housing Control', null=False, default=False,
